Step4 구현 완료 (#2666)
* :memo: README 4단계 작성

* :sparkles: feat(수동 관련 입출력 추가)

* :truck: refactor(package name 변경)

* :sparkles: feat(메뉴얼 로또 개수 로직 추가)

* :truck: refactor(package name 수정)

* :sparkles: feat(전체 구매가 수동일 때 처리)

* :recycle: refactor(로직 간결화 + empty 처리)

* :recycle: refactor(로직 간결화 + empty 처리)

* :fire: remove(필요없는 출력 삭제)

* :recycle: refactor(생성 로직 분리)

* :recycle: refactor(매직넘버 분리 및 생성 책임 분리)

* :recycle: refactor(Fake객체 생성 변경)

* :recycle: refactor(Bank 클래스 이름 변경)

* :bug: refactor(형변환 오류로 인한 버그 수정)

* :recycle: refactor(Bank 클래스 이름 변경)

* :recycle: refactor(변수 및 클래스 이름 변경)

* :recycle: refactor(test 변경)

* :recycle: refactor(machine 반복 수정)

* :recycle: refactor(필요없는 로직 제거)

* :recycle: refactor(Bank 클래스 이름 변경)

* :recycle: refactor(Test Code를 위한 Fake객체 위치 수정)

* :recycle: refactor(반복에 대한 것을 객체관리가 아닌 매개변수로 수정)

+++ src/main/java/lotto/domain/Rank.java
@@ -36,7 +36,7 @@ public enum Rank {
 
     static double yield(EnumMap<Rank, Integer> rankIntegerEnumMap, int purchaseMoney) {
         double sum = rankIntegerEnumMap.entrySet().stream()
-            .mapToDouble(rankIntegerEntry -> rankIntegerEntry.getValue() * rankIntegerEntry.getKey().reward)
+            .mapToDouble(rankIntegerEntry -> (double) rankIntegerEntry.getValue() * rankIntegerEntry.getKey().reward)
             .sum();
         return sum / purchaseMoney;
     }
